requirements.txt
boto3==1.2.2
PyYAML==3.11
